requirements.txt
fastapi
flask
psycopg2-binary